Restructured codec components
