--- components/rehash-codec-ffi-bindings/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "rehash-codec-ffi-bindings"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-libloading = "0.8.8"
-serde = { version = "1.0.219", features = ["derive"] }
