+++ components/rehash-codec/codec-ffi/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "codec-ffi"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
