+++ components/rehash-codec/codec-proc-macro/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "codec-proc-macro"
+version = "0.1.0"
+edition = "2024"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "2.0.104", features = ["full"] }
+proc-macro2 = "1.0.95"
