--- components/rehash-codec/src/lib.rs
@@ -1,4 +1,0 @@
-#[unsafe(no_mangle)]
-pub extern "C" fn add(left: u64, right: u64) -> u64 {
-    left + right
-}
