+++ components/rehash-frontend/Cargo.toml
@@ -11,7 +11,6 @@ wasm-bindgen = "0.2"
 js-sys = "0.3"
 wasm-bindgen-futures = "0.4.50"
 rehash-utils = { path = "../rehash-utils", features = ["tauri"] }
-rehash-codec-ffi-bindings = { path = "../rehash-codec-ffi-bindings" }
 serde = { version = "1.0.219", features = ["derive"] }
 serde-wasm-bindgen = "0.6.5"
