+++ rehash-desktop/Cargo.toml
@@ -26,4 +26,5 @@ tauri-plugin-fs = "2"
 serde = { version = "1.0.219", features = ["derive"] }
 tokio = "1.47.0"
 rehash-utils = { path = "../components/rehash-utils", features = [] }
-rehash-codec-ffi-bindings = { path = "../components/rehash-codec-ffi-bindings" }
+codec-ffi = { path = "../components/rehash-codec/codec-ffi" }
+libloading = "0.8.8"
